requirements.txt
absl-py==0.13.0
astunparse==1.6.3
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Flask==1.1.4
Flask-JWT-Extended==3.25.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.32.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
Pillow==8.2.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pytesseract==0.3.7
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.5
Werkzeug==1.0.1
wrapt==1.12.1
